Bump default deploy timeout to 90 minutes
We have started hitting the deploy timeout pretty regularly in the
HA jobs. Hopefully once we have composable roles, we can split the
jobs up to provide full coverage across multiple jobs while
lowering the footprint of the individual jobs. In the meantime, we
are at the mercy of the monolithic roles, and when more stuff is
added the deploy time will increase.

Change-Id: I4deeb5b4efff768d1e4f8d74b8eee055c151f4a5